[CHORE](foundation): Extract mock currents helpers
