+++ rust/foundation-api/src/routes/init.rs
@@ -1,4 +1,5 @@
-use super::whoami::whoami_and_authorize;
+use super::mock_currents::mock_currents_records;
+use super::{caller_token, whoami::whoami_and_authorize};
 use crate::{
     auth::AuthzAction, config::FoundationConfig, errors::ServerError, server::FoundationApiServer,
     wiki::WikiClientError,
@@ -19,7 +20,7 @@ use frontend_core::{
     },
     foundation::source_kind_for_collection_name,
 };
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use std::collections::HashMap;
 use uuid::Uuid;
 
@@ -336,16 +337,6 @@ impl ChromaError for FoundationInitError {
     }
 }
 
-const CHROMA_TOKEN_HEADER: &str = "x-chroma-token";
-
-fn chroma_token(headers: &HeaderMap) -> Result<&str, FoundationInitError> {
-    headers
-        .get(CHROMA_TOKEN_HEADER)
-        .and_then(|value| value.to_str().ok())
-        .filter(|token| !token.is_empty())
-        .ok_or(FoundationInitError::MissingToken)
-}
-
 async fn ensure_currents_collection(
     server: &FoundationApiServer,
     headers: &HeaderMap,
@@ -356,7 +347,7 @@ async fn ensure_currents_collection(
         .wiki_client
         .as_ref()
         .ok_or(FoundationInitError::WikiRouteUnavailable)?;
-    let token = chroma_token(headers)?;
+    let token = caller_token(headers).ok_or(FoundationInitError::MissingToken)?;
     let collection = wiki_client
         .get_collection_by_name(tenant, token, collection_name)
         .await?;
@@ -406,7 +397,7 @@ async fn maybe_seed_currents_collection(
         return Ok(());
     }
 
-    if chroma_token(headers).is_err() {
+    if caller_token(headers).is_none() {
         tracing::info!(
             collection_name = collection_name,
             "skipping currents mock seed because request carried no x-chroma-token"
@@ -415,102 +406,6 @@ async fn maybe_seed_currents_collection(
     }
 
     ensure_currents_collection(server, headers, tenant, collection_name).await
-}
-
-struct MockCurrentRecord {
-    id: String,
-    document: String,
-    metadata: Metadata,
-}
-
-#[derive(Debug, Deserialize)]
-struct MockCurrentFixture {
-    tilegroup_id: String,
-    label: String,
-    headline: String,
-    summary: String,
-    tiles: Vec<MockTileFixture>,
-}
-
-#[derive(Debug, Deserialize)]
-struct MockTileFixture {
-    slug: String,
-    title: String,
-    role: String,
-    blurb: String,
-}
-
-fn mock_currents_records() -> Vec<MockCurrentRecord> {
-    let fixture = include_str!("mock_currents.json");
-    let tilegroups: Vec<MockCurrentFixture> =
-        serde_json::from_str(fixture).expect("mock_currents.json must be valid");
-    tilegroups.into_iter().map(mock_current_record).collect()
-}
-
-fn mock_current_record(fixture: MockCurrentFixture) -> MockCurrentRecord {
-    let MockCurrentFixture {
-        tilegroup_id,
-        label,
-        headline,
-        summary,
-        tiles,
-    } = fixture;
-    let page_slugs: Vec<String> = tiles.iter().map(|tile| tile.slug.clone()).collect();
-    let tile_roles: Vec<String> = tiles.iter().map(|tile| tile.role.clone()).collect();
-    let mut metadata = Metadata::new();
-    metadata.insert(
-        "tilegroup_id".to_string(),
-        MetadataValue::Str(tilegroup_id.clone()),
-    );
-    metadata.insert("label".to_string(), MetadataValue::Str(label.clone()));
-    metadata.insert("headline".to_string(), MetadataValue::Str(headline.clone()));
-    metadata.insert("summary".to_string(), MetadataValue::Str(summary.clone()));
-    metadata.insert(
-        "tile_count".to_string(),
-        MetadataValue::Int(tiles.len() as i64),
-    );
-    metadata.insert(
-        "page_slugs".to_string(),
-        MetadataValue::StringArray(page_slugs),
-    );
-    metadata.insert(
-        "tile_roles".to_string(),
-        MetadataValue::StringArray(tile_roles),
-    );
-    for (idx, tile) in tiles.iter().enumerate() {
-        let key = format!("tile_{:02}_json", idx + 1);
-        let value = serde_json::json!({
-            "order": idx + 1,
-            "slug": tile.slug,
-            "title": tile.title,
-            "role": tile.role,
-            "blurb": tile.blurb,
-        });
-        metadata.insert(key, MetadataValue::Str(value.to_string()));
-    }
-
-    let tiles_document = tiles
-        .iter()
-        .enumerate()
-        .map(|(idx, tile)| {
-            format!(
-                "Tile {}\nSlug: {}\nTitle: {}\nRole: {}\nBlurb: {}",
-                idx + 1,
-                tile.slug,
-                tile.title,
-                tile.role,
-                tile.blurb
-            )
-        })
-        .collect::<Vec<_>>()
-        .join("\n\n");
-    let document = format!("{}: {}. {}\n\n{}", label, headline, summary, tiles_document);
-
-    MockCurrentRecord {
-        id: format!("tilegroup:{tilegroup_id}"),
-        document,
-        metadata,
-    }
 }
 
 async fn ensure_database(
